Move truss server out of common (#1146)
